IDEA: create rate package
create a flatter hierarchy of packages with a package for the
rate limit domain.

The proxy package makes sense to me :) The rules package could go into
the rate package IMHO. As these are rate limit rules as far as I understand.

so you could have a rule.go file in the rate pkg next to the rate.go file.